requirements.txt
fastapi
uvicorn
transformers
torch
Pillow
huggingface_hub
